requirements.txt
crewai
crewai-tools
python-dotenv
flask
crewai[google-genai]
PyYAML
gunicorn